Fix getClientIP to tkae 3rd from last X-F-F address
This is specific to our setup behind GCLB where the GCLB adds 2 ips to
xff.

Ideally this would not be necessary because lua API would have access to
the StreamInfo methods for getting the downstream address. We would
factorize the logic for determining which address is the legitimate
downstream client address in the X-F-F chain and delegate it to envoy &
its settings